MM-14958 Scroll to bottom when posting a new message (#2761)
