+++ NOTICE.txt
@@ -725,7 +725,7 @@ Common code (API client, Redux stores, logic, utility functions) for building a 
 
 * LICENSE: Apache-2.0
 
-Copyright 2015-present Mattermost, Inc. 
+Copyright 2015-present Mattermost, Inc.
 
                                  Apache License
                            Version 2.0, January 2004
@@ -1401,7 +1401,7 @@ SOFTWARE.
 
 This product contains 'react-native-doc-viewer' by Philipp Hecht.
 
-React Native Native Module Bridge Quicklock Document Viewer for IOS + Android supports pdf, png, jpg, xls, ppt, doc, docx, pptx, xlx + Video Player mp4 supported 
+React Native Native Module Bridge Quicklock Document Viewer for IOS + Android supports pdf, png, jpg, xls, ppt, doc, docx, pptx, xlx + Video Player mp4 supported
 
 * HOMEPAGE:
   * https://github.com/philipphecht/react-native-doc-viewer/blob/master/README.md
@@ -1859,41 +1859,6 @@ SOFTWARE.
 
 ---
 
-## react-native-recyclerview-list
-
-This product contains 'react-native-recyclerview-list' by GitHub user "godness84".
-
-A RecyclerView implementation for React Native
-
-* HOMEPAGE:
-  * https://github.com/godness84/react-native-recyclerview-list#readme
-
-* LICENSE: MIT
-
-The MIT License (MIT)
-
-Copyright (c) 2015 Marc Shilling
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
----
-
 ## react-native-safe-area
 
 This product contains 'react-native-safe-area' by Masayuki Iwai.
@@ -1942,16 +1907,16 @@ This package simplifies constructing the getItemLayout prop for react native Sec
 
 Copyright (c) 2017 Jan Soendermann
 
-Licensed under the Apache License, Version 2.0 (the "License"); 
-you may not use this file except in compliance with the License. 
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
 http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software 
-distributed under the License is distributed on an "AS IS" BASIS, 
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
-See the License for the specific language governing permissions and 
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
 limitations under the License.
 
 ---
