+++ android/app/src/main/java/com/mattermost/rnbeta/MainApplication.java
@@ -22,7 +22,6 @@ import com.masteratul.exceptionhandler.ReactNativeExceptionHandlerPackage;
 import com.RNFetchBlob.RNFetchBlobPackage;
 import com.gantix.JailMonkey.JailMonkeyPackage;
 import io.tradle.react.LocalAuthPackage;
-import com.github.godness84.RNRecyclerViewList.RNRecyclerviewListPackage;
 import com.reactnativecommunity.webview.RNCWebViewPackage;
 import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 
@@ -84,7 +83,6 @@ public class MainApplication extends NavigationApplication implements INotificat
             new SharePackage(this),
             new KeychainPackage(),
             new InitializationPackage(this),
-            new RNRecyclerviewListPackage(),
             new RNCWebViewPackage(),
             new RNGestureHandlerPackage()
     );
