+++ app/components/post_list/post_list.js
@@ -3,20 +3,36 @@
 
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
+import {FlatList, StyleSheet} from 'react-native';
 
+import EventEmitter from 'mattermost-redux/utils/event_emitter';
 import * as PostListUtils from 'mattermost-redux/utils/post_list';
 
 import CombinedUserActivityPost from 'app/components/combined_user_activity_post';
 import Post from 'app/components/post';
-import {DeepLinkTypes} from 'app/constants';
+import {DeepLinkTypes, ListTypes} from 'app/constants';
 import mattermostManaged from 'app/mattermost_managed';
+import {makeExtraData} from 'app/utils/list_view';
 import {changeOpacity} from 'app/utils/theme';
 import {matchDeepLink} from 'app/utils/url';
 
 import DateHeader from './date_header';
 import NewMessagesDivider from './new_messages_divider';
 
-export default class PostListBase extends PureComponent {
+const INITIAL_BATCH_TO_RENDER = 15;
+const SCROLL_UP_MULTIPLIER = 3.5;
+const SCROLL_POSITION_CONFIG = {
+
+    // To avoid scrolling the list when new messages arrives
+    // if the user is not at the bottom
+    minIndexForVisible: 0,
+
+    // If the user is at the bottom or 60px from the bottom
+    // auto scroll show the new message
+    autoscrollToTopThreshold: 60,
+};
+
+export default class PostList extends PureComponent {
     static propTypes = {
         actions: PropTypes.shape({
             handleSelectChannelByName: PropTypes.func.isRequired,
@@ -58,6 +74,30 @@ export default class PostListBase extends PureComponent {
         siteURL: '',
     };
 
+    constructor(props) {
+        super(props);
+
+        this.hasDoneInitialScroll = false;
+        this.contentOffsetY = 0;
+        this.makeExtraData = makeExtraData();
+        this.flatListRef = React.createRef();
+
+        this.state = {
+            postListHeight: 0,
+        };
+    }
+
+    componentDidMount() {
+        EventEmitter.on('scroll-to-bottom', this.scrollToBottom);
+    }
+
+    componentWillReceiveProps(nextProps) {
+        if (this.props.channelId !== nextProps.channelId) {
+            this.contentOffsetY = 0;
+            this.hasDoneInitialScroll = false;
+        }
+    }
+
     componentDidUpdate() {
         if (this.props.deepLinkURL) {
             this.handleDeepLink(this.props.deepLinkURL);
@@ -65,10 +105,23 @@ export default class PostListBase extends PureComponent {
         }
     }
 
+    componentWillUnmount() {
+        EventEmitter.off('scroll-to-bottom', this.scrollToBottom);
+    }
+
     handleClosePermalink = () => {
         const {actions} = this.props;
         actions.selectFocusedPostId('');
         this.showingPermalink = false;
+    };
+
+    handleContentSizeChange = (contentWidth, contentHeight) => {
+        this.scrollToInitialIndexIfNeeded(contentWidth, contentHeight);
+
+        if (this.state.postListHeight && contentHeight < this.state.postListHeight && this.props.extraData) {
+            // We still have less than 1 screen of posts loaded with more to get, so load more
+            this.props.onLoadMoreUp();
+        }
     };
 
     handleDeepLink = (url) => {
@@ -81,6 +134,22 @@ export default class PostListBase extends PureComponent {
             } else if (match.type === DeepLinkTypes.PERMALINK) {
                 this.handlePermalinkPress(match.postId, match.teamName);
             }
+        }
+    };
+
+    handleLayout = (event) => {
+        const {height} = event.nativeEvent.layout;
+        this.setState({postListHeight: height});
+    };
+
+    handlePermalinkPress = (postId, teamName) => {
+        const {actions, onPermalinkPress} = this.props;
+
+        if (onPermalinkPress) {
+            onPermalinkPress(postId, true);
+        } else {
+            actions.loadChannelsByTeamName(teamName);
+            this.showPermalinkView(postId);
         }
     };
 
@@ -100,15 +169,29 @@ export default class PostListBase extends PureComponent {
         }
     };
 
-    handlePermalinkPress = (postId, teamName) => {
-        const {actions, onPermalinkPress} = this.props;
+    handleScroll = (event) => {
+        const pageOffsetY = event.nativeEvent.contentOffset.y;
+        if (pageOffsetY > 0) {
+            const contentHeight = event.nativeEvent.contentSize.height;
+            const direction = (this.contentOffsetY < pageOffsetY) ?
+                ListTypes.VISIBILITY_SCROLL_UP :
+                ListTypes.VISIBILITY_SCROLL_DOWN;
 
-        if (onPermalinkPress) {
-            onPermalinkPress(postId, true);
-        } else {
-            actions.loadChannelsByTeamName(teamName);
-            this.showPermalinkView(postId);
+            this.contentOffsetY = pageOffsetY;
+            if (
+                direction === ListTypes.VISIBILITY_SCROLL_UP &&
+                (contentHeight - pageOffsetY) < (this.state.postListHeight * SCROLL_UP_MULTIPLIER)
+            ) {
+                this.props.onLoadMoreUp();
+            }
         }
+    };
+
+    handleScrollToIndexFailed = () => {
+        requestAnimationFrame(() => {
+            this.hasDoneInitialScroll = false;
+            this.scrollToInitialIndexIfNeeded(1, 1);
+        });
     };
 
     keyExtractor = (item) => {
@@ -178,6 +261,27 @@ export default class PostListBase extends PureComponent {
         );
     };
 
+    scrollToBottom = () => {
+        this.flatListRef.current.scrollToOffset({offset: 0, animated: true});
+    };
+
+    scrollToInitialIndexIfNeeded = (width, height) => {
+        if (
+            width > 0 &&
+            height > 0 &&
+            this.props.initialIndex > 0 &&
+            !this.hasDoneInitialScroll
+        ) {
+            this.flatListRef.current.scrollToIndex({
+                animated: false,
+                index: this.props.initialIndex,
+                viewOffset: 50,
+                viewPosition: 0.5,
+            });
+            this.hasDoneInitialScroll = true;
+        }
+    };
+
     showPermalinkView = (postId) => {
         const {actions, navigator} = this.props;
 
@@ -204,5 +308,51 @@ export default class PostListBase extends PureComponent {
             navigator.showModal(options);
         }
     };
+
+    render() {
+        const {
+            channelId,
+            highlightPostId,
+            postIds,
+            refreshing,
+        } = this.props;
+
+        const refreshControl = {refreshing};
+
+        if (channelId) {
+            refreshControl.onRefresh = this.handleRefresh;
+        }
+
+        const hasPostsKey = postIds.length ? 'true' : 'false';
+
+        return (
+            <FlatList
+                key={`recyclerFor-${channelId}-${hasPostsKey}`}
+                ref={this.flatListRef}
+                contentContainerStyle={styles.postListContent}
+                data={postIds}
+                extraData={this.makeExtraData(channelId, highlightPostId, this.props.extraData)}
+                initialNumToRender={INITIAL_BATCH_TO_RENDER}
+                inverted={true}
+                keyExtractor={this.keyExtractor}
+                ListFooterComponent={this.props.renderFooter}
+                maintainVisibleContentPosition={SCROLL_POSITION_CONFIG}
+                maxToRenderPerBatch={INITIAL_BATCH_TO_RENDER + 1}
+                onContentSizeChange={this.handleContentSizeChange}
+                onLayout={this.handleLayout}
+                onScroll={this.handleScroll}
+                onScrollToIndexFailed={this.handleScrollToIndexFailed}
+                removeClippedSubviews={true}
+                renderItem={this.renderItem}
+                scrollEventThrottle={60}
+                {...refreshControl}
+            />
+        );
+    }
 }
 
+const styles = StyleSheet.create({
+    postListContent: {
+        paddingTop: 5,
+    },
+});
