+++ app/components/post_textbox/post_textbox.js
@@ -383,6 +383,8 @@ export default class PostTextbox extends PureComponent {
 
                 this.setState(nextState, callback);
             }
+
+            EventEmitter.emit('scroll-to-bottom');
         }
     };
